refractor analytics, move some functions to extension-analytics
