+++ src/services.js
@@ -28,6 +28,7 @@ import {
   getElementServiceIfAvailable,
   getElementServiceIfAvailableForDoc,
 } from './element-service';
+import {getAmpdoc} from './service';
 
 
 /**
@@ -332,4 +333,32 @@ export function viewportForDoc(nodeOrDoc) {
  */
 export function xhrFor(window) {
   return /** @type {!./service/xhr-impl.Xhr} */ (getService(window, 'xhr'));
+}
+
+/**
+ * @param {!Node|!./service/ampdoc-impl.AmpDoc} nodeOrDoc
+ * @param {boolean=} loadAnalytics
+ * @return {!Promise<!../extensions/amp-analytics/0.1/instrumentation.InstrumentationService>}
+ */
+export function analyticsForDoc(nodeOrDoc, loadAnalytics = false) {
+  if (loadAnalytics) {
+    // Get Extensions service and force load analytics extension.
+    const ampdoc = getAmpdoc(nodeOrDoc);
+    extensionsFor(ampdoc.win)./*OK*/loadExtension('amp-analytics');
+  }
+  return (/** @type {!Promise<
+            !../extensions/amp-analytics/0.1/instrumentation.InstrumentationService
+          >} */ (getElementServiceForDoc(
+              nodeOrDoc, 'amp-analytics-instrumentation', 'amp-analytics')));
+}
+
+/**
+ * @param {!Node|!./service/ampdoc-impl.AmpDoc} nodeOrDoc
+ * @return {!Promise<?../extensions/amp-analytics/0.1/instrumentation.InstrumentationService>}
+ */
+export function analyticsForDocOrNull(nodeOrDoc) {
+  return (/** @type {!Promise<
+            ?../extensions/amp-analytics/0.1/instrumentation.InstrumentationService
+          >} */ (getElementServiceIfAvailableForDoc(
+              nodeOrDoc, 'amp-analytics-instrumentation', 'amp-analytics')));
 }
